feat: Use FastMCP for boilerplate, much cleaner code
