+++ pyproject.toml
@@ -5,6 +5,7 @@ description = "Connect Claude conversations with AnkiConnect via MCP to make spa
 readme = "README.md"
 requires-python = ">=3.11"
 dependencies = [
+    "fastmcp>=0.4.1",
     "httpx>=0.28.1",
     "mcp>=1.1.0",
     "pytest-asyncio>=0.24.0",
